[NOOP] spring-boot/gradle-plugin use platform API
Modified/added gradle files are commented so this changeset should be
functionally equivalent to parent commit.

Failing attempt to configure the new spring-boot-gradle-plugin to use
Gradle's native Platform API instead of Spring's dependency management
plugin.

According to the Internet this should be possible but currently
does not seem worth the aggravation given that spring-boot's issues page
on GitHub indicates that a coming release will document this officially.